+++ src/files/XmlFile.ts
@@ -559,6 +559,8 @@ export class XmlFile {
                 let missingImports = this.getMissingImportsForTranspile()
                     //change the file extension to .brs since they will be transpiled
                     .map(x => x.replace('.bs', '.brs'));
+                //always include the bslib file
+                missingImports.push('source/bslib.brs');
 
                 for (let missingImport of missingImports) {
                     let scriptTag = `<script type="text/brightscript" uri="${util.getRokuPkgPath(missingImport)}" />`;
@@ -569,6 +571,8 @@ export class XmlFile {
                         new SourceNode(1, 0, this.pathAbsolute, indent + scriptTag)
                     );
                 }
+            } else {
+                //we couldn't find the closing component tag....so maybe there's something wrong? or this isn't actually a component?
             }
 
             //convert .bs extensions to .brs